+++ harmonyAsSwitches.js
@@ -84,6 +84,7 @@ HarmonyPlatformAsSwitches.prototype = {
 
     this.harmonyBase.getDevicesAccessories(this, data);
     this.harmonyBase.getSequencesAccessories(this, data);
+    this.harmonyBase.getHomeControlsAccessories(this);
 
     //first refresh