+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.6.0
+
+- [NEW] - support for Exposing Home Control buttons #67
+
 ## 0.5.5
 
 - [BUG] - showTurnOffActivity doesn’t work properly when inverterted #107